Render

+++ android-final-project/app/src/main/java/com/mybaby/android_final_project/model/MeasurePerMonth.java
@@ -25,7 +25,7 @@ public class MeasurePerMonth implements Comparable<MeasurePerMonth>{
         return month;
     }
 
-    public float getMeasure() {
+    public double getMeasure() {
         return measure;
     }
 }
